refactor(Counter): 删除 count-selector 组件
count-selector 是 counter 更名前的组件，现在已无用

BREAKING CHANGE: 删除 CountSelector 组件
