--- examples/pages/components/shopping/pages/count-selector/index.wxml
@@ -1,5 +1,0 @@
-<content-title name="CountSelector" describe="数量选择器">
-  <content-card name="{{item.title}}" wx:for="{{countConfig}}" wx:key="item">
-    <l-count-selector l-class="count" wx:for="{{item.config}}" wx:for-item="count" wx:key="count" min="{{count.min}}" count="{{count.count}}" max="{{count.max}}" step="{{count.step}}" disabled="{{count.disabled}}" bind:linout="onOut"/>
-  </content-card>
-</content-title>
