+++ examples/pages/components/shopping/pages/counter/index.wxss
