--- src/count-selector/index.json
@@ -1,6 +1,0 @@
-{
-  "component": true,
-  "usingComponents": {
-    "l-icon":"../icon/index"
-  }
-}
