--- src/count-selector/index.less
@@ -1,68 +1,0 @@
-.container-count {
-  display       : flex;
-  flex-direction: row;
-  width         : 170rpx;
-  height        : 56rpx;
-}
-
-.symbol {
-  height         : 100%;
-  width          : 56rpx;
-  font-size      : 28rpx;
-  color          : #596C8E;
-  display        : flex;
-  align-items    : center;
-  justify-content: center;
-}
-
-.disabled {
-  background-color: #f3f3f3;
-  color           : #c4c9d2;
-}
-
-.abled {
-  background-color: #ecf1f8;
-  color           : #596c8e;
-}
-
-.count {
-  height       : 100%;
-  flex         : 1;
-  min-height   : 56rpx;
-  line-height  : 56rpx;
-  font-size    : 24rpx;
-  color        : #596c8e;
-  background   : #f6f7f9;
-  text-align   : center;
-  overflow     : hidden;
-  text-overflow: ellipsis;
-  white-space  : nowrap;
-}
-
-@font-face {
-  font-family: "iconfont";
-  src        : url('data:application/x-font-woff2;charset=utf-8;base64,d09GMgABAAAAAALEAAsAAAAABqwAAAJ3AAEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAHEIGVgCDBgqBFIEWATYCJAMMCwgABCAFhG0HNxvdBcgekiSBqhRCJKIogADMEYRr5GySu0dQoBAloETFRVKmqpWoK1wVG1th2Ij732n7N5fsqTh2xyneKfa2N3/anZKGvYXFohGGrUFIEA6JogD/fZ7L6U2gA7nGtyynNW3s8fyoF2AcUEBjbYqsRALxFtlF3ImLOE6g3ZheoYOxmRVIKtC4QLww9QYkCxFFMfKtQt2wtIjXKq3pZfrEq+j78d9qJEmqDE09fTTahoFf6deHfNxwgxOiBHl1hYx5oBC3GjMXIsLgItpN0e7yWhHSXPEmkcq31Dv94yWihmo7wbQcS/yiMoJf9yskkEFdzE0Diwo7eF2IkexeNr35cvtSjmQ/fjt5zeTy++93z1I4x8LHT3u9VVdH6SHr0r8+OjOTtCopyONe0uFsIlZDiaMZyTdAmzFfZzrm0+EOQPOcP/IeANGP4rNA8KN0+V8rbPwlMwX83P7UK0Wa5wJwS+sZvMEfUwMbiq9F1lxcFVWFvmwE4ErICtOvARTo9zPeMgROEVoL2hG+FqMeslbjZGHnUemwjlqrbbSbM766wwAuIkobZl1ICL0ekXT7jqzXF1nYb1SG/aLWGxG0O4uBLTtMhpne59gQ2CbNHWK2mEeJXaTj+iXUN5wGz2sjfg15YMaJJqvl/Dh6yOeYEWzqFSEooZy5ZEw6DR2HEZ8zC1tCNoTwq4pCm94kt5gLafs41CCgNqJpB2FqYTyUtxbSlc8vQboNjgbeIejJr0FcwOyc0MhUADlu8kCCe3klsElXIQgUQXGMixiTBiGHgyH85kEW1CLIjBEpX5WCm1GoSV7f4n7fNmiH9syRIkeR77bbiWdmwzMaDAAA') format('woff2')
-}
-
-.l-icon-add:before {
-  content: "\e602";
-}
-
-.l-icon-reduce:before {
-  content: "\e69c";
-}
-
-.l-icon {
-  font-family            : "iconfont";
-  font-style             : normal;
-  -webkit-font-smoothing : antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-.l-icon::before {
-  display: inline-flex;
-}
-
-.count-hover{
-  opacity: .8;
-}
